updated install script
Added a check if there is an existing configuration. If so it won't override it.

+++ install.sh
@@ -10,9 +10,10 @@
 giturl="https://github.com/mnul/debian-autoshutdown/archive/refs/heads/master.zip"
 TMPDIR=/tmp/debian-autoshutdown
 ZIPNAME="debian-autoshutdown.zip"
+EXISTCONF=0
+
 
 #Install-Script starts here
-
 
 
 if [ $USER != 'root' ] #check if script is run as root, if not exit
@@ -39,6 +40,7 @@ if [[ $(pwd) != $TMPDIR ]]
 fi
 
 
+# check if curl or wget are available and donwload the files. If not error -> exit
 if [ -x "$(which wget)" ] ;
   then
     wget -O $ZIPNAME $giturl
@@ -51,6 +53,7 @@ else
   exit
 fi
 
+# unzip the files, if unzip doesnt exist exit
 if [ -x "$(which unzip)" ];
   then
     unzip -q $ZIPNAME
@@ -59,33 +62,56 @@ if [ -x "$(which unzip)" ];
     exit
 fi
 
+# copy files to their destination
 cd debian-autoshutdown-master
 
 sudo cp -R etc/* /etc/
 sudo cp -R usr/* /usr/
 sudo cp -R lib/* /lib/
 sudo cp autoshutdown.service /etc/systemd/system/autoshutdown.service
-sudo cp /etc/autoshutdown.default /etc/autoshutdown.conf
+
+# check if there is an existing autoshutdown.conf if so the skip copying.
+if [ -f "/etc/autoshutdown.conf"];
+  then
+    echo "" && echo "" && echo ""
+    echo "configuration file exists, skipping creation of default configuration."
+    $EXISTCONF=1
+  else  
+    sudo cp /etc/autoshutdown.default /etc/autoshutdown.conf
+    $EXISTCONF=0
+fi
 
 echo ""
 echo ""
 echo "all files copied to their location!"
 echo ""
 
-
+# (re)start the service
 sudo systemctl daemon-reload
 sudo systemctl start autoshutdown
 sudo systemctl enable autoshutdown
 sudo systemctl daemon-reload
 
+# Script copied. Let user know how to proceed
+
+echo "***********************  DONE! ***********************************************"
 echo "" && echo "" && echo ""
 echo "" && echo "" && echo ""
-echo "Autoshutdown is now active with its default configuration"
-echo ""
-echo "to finish and configure the service you must edit its configuration file (in this case using the nano editor:"
-echo ""
-echo "->    sudo nano /etc/autoshutdown.conf"
-echo ""
+if [ $EXISTCONF -eq 1] 
+  then
+    echo "Autoshutdown is updated to the latest version."
+    echo "" && echo ""
+    echo "please compare your configured /etc/autoshutdown.conf and the default /etc/autoshutdown.default to see if configs have changed."
+    echo ""
+    echo "If so either update your current configuration to match the new default or start over by overwriting the current configuration with the new default"
+  else
+    echo "Autoshutdown is now active with its default configuration"
+    echo ""
+    echo "to finish and configure the service you must edit its configuration file (in this case using the nano editor:"
+    echo ""
+    echo "->    sudo nano /etc/autoshutdown.conf"
+    echo ""
+fi
 echo "Finally restart autoshutdown to enable your changes"
 echo ""
 echo "->    sudo systemctl restart autoshutdown"
